fix: AddPrinterOptions broke in 0.3.1
switching the binding from o.OutputFormat to o.DefaultOutputFormat
meant that we were no longer ensuring that the o.OutputFormat pointer
was non-nil, meaning that when it was nil the addObjectPrinterFlags()
and addListPrinterFlags() methods were short-circuting before they
added the output/o flag.

updating o.OutputFormat from a *string to a string means that the
memory location always exists and we can safely bind to it, using
the ActiveOutputFormat() (which will resolve safely to the value
of o.DefaultOutputFormat or "text" as set when the flags are added)
as the default value for the flag.